+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Changed
 
+- The OKR period selection tabs have been moved into a drop down menu in the top
+  bar matching the period selector on the KPI page.
 - Improved OKR overview page styling.
 - Improved tooltip styling.